Add shave_crc_from_ipv4 to remove eth crc field
The crc checksum from the ethernet frame is ignored by the smoltcp
EthernetFrame struct (See: https://github.com/m-labs/smoltcp/issues/77).
This new function checks if the crc frame is left on the end of the ipv4
frame and reduces the slice size accordingly. This fixes any out of
bounds errors when handline fragmentation from when the ipv4 length is assumed based on the length
of the slice rather than the header value.